# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
...
  '/api/films/public': 
    get:
      tags:
        - films/public
      summary: Retrieve the public films
      description: >-
        The public films (i.e., the films that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPublicFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
                  next:
                    type: string
                    format: uri
              examples:
                Example:
                  value:
                    totalPages: 3
                    currentPage: 1
                    totalItems: 6
                    films:
                      - id: 1
                        title: "Your Name"
                        owner: 1
                        private: false
                        watchDate: "2021-10-03"
                        rating: 9
                        favorite: true
                        self: "/api/films/public/1"
                      - id: 2
                        title: "Heaven's Feel"
                        owner: 2
                        private: false
                        self: "/api/films/public/2"
                    next": "/api/films/public?pageNo=2"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
...
  '/api/films/public/{filmId}/reviews/{reviewId}/drafts':
    post:
      tags:
        - 'films/public/{filmId}/reviews/{reviewId}/drafts'
      security:
        - cookieAuth: []
      summary: Create a draft for a film review
      description: >-
        A for the review with ID reviewId of the film with ID filmId is created. The draft is created by the user with ID reviewerId, and contains a proposed review and a proposed rating. This operation can be performed only by one of the reviewers for whom the review was issued.
      operationId: createDraft
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
        - name: reviewId
          in: path
          description: ID of the review to which the draft is associated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: the new film review draft, including the author, proposed review and proposed rating
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Draft'
            examples:
              Example:
                value:
                  reviewerId: 3
                  proposedRating: 5
                  proposedReview: "Quite engaging movie."
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
              examples:
                Example:
                  value:
                    draftId: 5
                    reviewId: 8
                    reviewerId: 3
                    proposedRating: 5
                    proposedReview: "Quite engaging movie."
                    open: true
                    self: "/api/films/public/6/reviews/8/drafts/5"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                invalid parameters:
                  value:
                    errors:
                      - param: Server
                        msg: Invalid filmId or reviewId, they must be a number.
                missing parameters:
                  value:
                    errors:
                      - param: Server
                        msg: filmId or reviewId parameters missing
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example 1:
                  value:
                    errors:
                      - param: Server
                        msg: The user is not one of the assigned reviewers.
        '409':
          description: Conflict (e.g., wrong film id or review id)
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example 1:
                  value:
                    errors:
                      - param: Server
                        msg: The reviewerId field of the draft object is different from the id of the user requesting the operation.
                Example 2:
                  value:
                    errors:
                      - param: Server
                        msg: There is already an open draft for this review. 
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
    get:
      tags:
        - 'films/public/{filmId}/reviews/{reviewId}/drafts'
      summary: Retrieve the drafts that has been issued/completed by a group with review ID reviewId for a film with ID filmId
      description: >-
        The drafts of the film with ID filmID issued to the review with ID reviewId is retrieved. This operation requires authentication and only the owner of the film or the reviewers for whom the review was issued can view the drafts. 
      operationId: getAllDrafts
      parameters:
        - name: filmId
          in: path
          description: ID of the film 
          required: true
          schema:
            type: integer
            format: int64
        - name: reviewId
          in: path
          description: ID of the review to which the draft is associated
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                    $ref: '#/components/schemas/Draft'
              examples:
                Example:
                  value:
                    totalPages: 1
                    currentPage: 1
                    totalItems: 2
                    drafts:
                      - draftId: 2
                        reviewId: 6
                        reviewerId: 5
                        proposedRating: 5
                        proposedReview: "A heartwarming and touching story that will leave you feeling uplifted and inspired. A must-see for all ages."
                        open: false
                        self: "/api/films/public/6/reviews/6/drafts/2"
                      - draftId: 3
                        reviewId: 6
                        reviewerId: 2
                        proposedRating: 2
                        proposedReview: "A dreadful movie with little original ideas in it."
                        open: true
                        self: "/api/films/public/6/reviews/6/drafts/3"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                invalid parameters:
                  value:
                    errors:
                      - param: Server
                        msg: Invalid filmId or reviewId, they must be a number.
                missing parameters:
                  value:
                    errors:
                      - param: Server
                        msg: filmId or reviewId parameters missing
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example 1:
                  value:
                    errors:
                      - param: Server
                        msg: The user is not the owner of the film nor is it one of the assigned reviewers.
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example 1:
                  value:
                    errors:
                      - param: Server
                        msg: There's no review with ID reviewId associated to the film with ID filmId
                Example 2:
                  value:
                    errors:
                      - param: Server
                        msg: There are no drafts for the review with ID reviewId associated to a film with ID filmId 
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
  '/api/films/public/{filmId}/reviews/{reviewId}/drafts/{draftId}':
    get:
      tags:
        - 'films/public/{filmId}/reviews/{reviewId}/drafts/{draftId}'
      summary: Retrieve a draft associated to the review with ID reviewId, issued for the film with ID filmId.
      description: >-
        The draft of the review with ID reviewId, issued for the film with ID filmID is retrieved. This operation requires authentication. Only the owner of the film or the reviewers for whom the review was issued can view the draft. 
      operationId: getSingleDraft
      parameters:
        - name: draftId
          in: path
          description: ID of draft 
          required: true
          schema:
            type: integer
            format: int64
        - name: reviewId
          in: path
          description: ID of the review associated with the requested draft
          required: true
          schema:
            type: integer
            format: int64
        - name: filmId
          in: path
          description: ID of the film associated to the review 
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
              examples:
                Example:
                  value:
                    draftId: 2
                    reviewId: 6
                    reviewerId: 5
                    proposedRating: 5
                    proposedReview: "A heartwarming and touching story that will leave you feeling uplifted and inspired. A must-see for all ages."
                    open: false
                    self: "/api/films/public/6/reviews/6/drafts/2"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                invalid parameters:
                  value:
                    errors:
                      - param: Server
                        msg: Invalid filmId, reviewId or draftId, they must be a number.
                missing parameters:
                  value:
                    errors:
                      - param: Server
                        msg: filmId, reviewId and/or draftId parameters missing
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example 1:
                  value:
                    errors:
                      - param: Server
                        msg: The user is not the owner of the film nor is it one of the assigned reviewers.
        '404':
          description: Not Found
          content:  
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The draft does not exist.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
...
components:
  schemas:
...
    Film:
      description: schema for Film data structures
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON data structure
            describing the film
        id:
          type: integer
          description: unique identifier for the film data structure
        title:
          type: string
          description: textual title of the film
        owner:
          type: integer
          description: unique identifier of the the film data structure owner
        private:
          type: boolean
          description: true if only the owner of the film data structure can access it, false otherwise
          default: 'true'
        watchDate:
          type: string
          format: date
          description: the date when the film was watched by the owner
        rating:
          type: integer
          description: the rating the owner has given to the film
          minimum: 1
          maximum: 10
        favorite:
          type: boolean
          description: 'true if the film is among the favourite
            ones of the owner, false otherwise '
          default: 'false'
        self: 
          type: string
          format: uri
      required:
      - title
      - private
...
    Draft:
      description: JSON Schema for Review Draft data structures
      type: object
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON Schema to use for the validation of the data structure
        draftId:
          type: integer
          description: unique identifier of the draft associated to a review identified by reviewId
        reviewId:
          type: integer
          description: unique identifier of the review to which the draft is associated
        proposedRating:
          type: integer
          description: the proposed rating of the review
          minimum: 1
          maximum: 10
        proposedReview:
          type: string
          description: a textual description of the proposed review
          maxLength: 1000
        reviewerId:
          type: integer
          description: the id of the author of the review draft
        open:
          type: boolean
          description: a boolean property set to true upon draft creation. It is set to
            false if all review co-assignees have expressed their agreement or disagreement.
          default: 'true'
        self: 
          type: string
          format: uri
      required:
      - proposedRating
      - proposedReview
      - reviewerId
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid